errors fixed

+++ app/Http/Controllers/TagController.php
@@ -10,6 +10,7 @@ use Illuminate\Support\Str;
 use App\Models\Account;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Schema;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Auth;
 
 
@@ -22,31 +23,31 @@ class TagController extends Controller
         $userId = $user->id;
     
 
-        return $request;
-        // $validator = Validator::make($request->all(), [
-        //     'category' => 'required|in:pet,kid,luggage',
-        //     'name' => $request->category === 'luggage' ? "nullable" : "required|string",
-        //     'ownerName' => $request->category === "kid" ? "nullable" : "required",
-        //     'gender' => $request->category === 'luggage' ? "nullable" : "required|in:male,female",
-        //     'age' => $request->category === 'luggage' ? "nullable" : "required",
-        //     'medicalIssue' => $request->category === 'luggage' ? "nullable" : "required",
-        //     'height' => $request->category === 'kid' ? "required" : "nullable",
-        //     'weight' => $request->category === 'pet' ? "required" : "nullable",
-        //     'color' => $request->category === 'pet' ? "required" : "nullable",
-        //     'dressColor' => $request->category === 'kid' ? "required" : "nullable",
-        //     'vetDetail' => $request->category === 'pet' ? "required" : "nullable",
-        //     'doctorDetail' => $request->category === 'kid' ? 'required' :"nullable",
-        //     'brand' => $request->category === 'luggage' ? "required" : "nullable",
-        //     'luggageType' => $request->category === 'luggage' ? "required" : "nullable",
-        //     'reward' => "required|string",
-        //     'mobileNumber' => "required|string",
-        //     'mobileNumber2' => "string|nullable",
-        //     'contactEmail' => "required|email",
-        //     'address' => "required|string",
-        //     'note' => "string|nullable",
-        //     'images' => 'nullable|array'
+      
+        $validator = Validator::make($request->all(), [
+            'category' => 'required|in:pet,kid,luggage',
+            'name' => $request->category === 'luggage' ? "nullable" : "required|string",
+            'ownerName' => $request->category === "kid" ? "nullable" : "required",
+            'gender' => $request->category === 'luggage' ? "nullable" : "required|in:male,female",
+            'age' => $request->category === 'luggage' ? "nullable" : "required",
+            'medicalIssue' => $request->category === 'luggage' ? "nullable" : "required",
+            'height' => $request->category === 'kid' ? "required" : "nullable",
+            'weight' => $request->category === 'pet' ? "required" : "nullable",
+            'color' => $request->category === 'pet' ? "required" : "nullable",
+            'dressColor' => $request->category === 'kid' ? "required" : "nullable",
+            'vetDetail' => $request->category === 'pet' ? "required" : "nullable",
+            'doctorDetail' => $request->category === 'kid' ? 'required' :"nullable",
+            'brand' => $request->category === 'luggage' ? "required" : "nullable",
+            'luggageType' => $request->category === 'luggage' ? "required" : "nullable",
+            'reward' => "required|string",
+            'mobileNumber' => "required|string",
+            'mobileNumber2' => "string|nullable",
+            'contactEmail' => "required|email",
+            'address' => "required|string",
+            'note' => "string|nullable",
+            'images' => 'nullable|array'
 
-        // ]);
+        ]);
 
         if ($validator->fails()) {
 
@@ -55,65 +56,65 @@ class TagController extends Controller
         }
     
 
-        // $tag = new Tag();
+        $tag = new Tag();
 
-        // $data = $request->all();
-        // $columns = Schema::getColumnListing('tags_category');
+        $data = $request->all();
+        $columns = Schema::getColumnListing('tags_category');
 
-        // foreach ($data as $key => $value) {
+        foreach ($data as $key => $value) {
 
-        //     if (in_array($key, $columns)) {
+            if (in_array($key, $columns)) {
 
-        //         $tag->$key = $value;
-        //     }
-        // }
-        // $tag->userId = $userId;
-        // // dd($tag);
-        // $tag->save();
+                $tag->$key = $value;
+            }
+        }
+        $tag->userId = $userId;
+        // dd($tag);
+        $tag->save();
 
         // Saving the Uploaded Images and their path in Database //
 
         // if ($request->hasFile('images') && is_array($request->file('images')) && count($request->file('images')) > 0) {
-        // if ($request->images) {
+        if ($request->images) {
 
-        //     // foreach ($request->images as $image) {
-        //     //     // Validate each image
-        //     //     $validator = Validator::make(
-        //     //         ['image' => $image],
-        //     //         ['image' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048'] // Adjust the validation rules as needed
-        //     //     );
+            // foreach ($request->images as $image) {
+            //     // Validate each image
+            //     $validator = Validator::make(
+            //         ['image' => $image],
+            //         ['image' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048'] // Adjust the validation rules as needed
+            //     );
 
-        //     //     if ($validator->fails()) {
-        //     //         // Handle validation errors
-        //     //         return response()->json(['status' => false, 'message' => "image validation fails"], 422);
-        //     //     }
-        //     // }
+            //     if ($validator->fails()) {
+            //         // Handle validation errors
+            //         return response()->json(['status' => false, 'message' => "image validation fails"], 422);
+            //     }
+            // }
 
-        //     // All images are valid, proceed to store them
-        //     foreach ($request->images as $image) {
+            // All images are valid, proceed to store them
+            foreach ($request->images as $image) {
 
-        //         $fileName = Str::random(8) . '_' . time() . '.' . $image->getClientOriginalExtension();
-        //         $path = $image->storeAs('uploads', $fileName, 'public');
+                $fileName = Str::random(8) . '_' . time() . '.' . $image->getClientOriginalExtension();
+                $path = $image->storeAs('uploads', $fileName, 'public');
 
-        //         // Create a database record for each stored image
-        //         Image::create([
-        //             'path' => $path,
-        //             'tagId' => $tag->id,
-        //         ]);
-        //     }
-        // } else {
-        //     // Handle if no images were uploaded
-        //     return response()->json(['status' => false, 'message' => 'No images uploaded'], 422);
-        // }
+                // Create a database record for each stored image
+                Image::create([
+                    'path' => $path,
+                    'tagId' => $tag->id,
+                ]);
+            }
+        } else {
+            // Handle if no images were uploaded
+            return response()->json(['status' => false, 'message' => 'No images uploaded'], 422);
+        }
 
 
-        // return response()->json(["status" => true, "message" => "Data inserted Successfully"]);
+        return response()->json(["status" => true, "message" => "Data inserted Successfully"]);
     }
 
 
     public function getTagsByCategory($cat)
     {
-        if ($cat != "pet" && $cat != "kid" && $cat != "luggage") {
+        if ($cat != "pet" && $cat != "kisd" && $cat != "luggage") {
             return response()->json(["status" => false, "message" => "Category does not Exist"], 404);
         }
 
@@ -203,58 +204,7 @@ class TagController extends Controller
     }
 
 
-    // public function getAllTags()
-    // {
-    //     $data = Tag::all();
-
-    //     $kidArray = [];
-    //     $petArray = [];
-    //     $luggageArray = [];
-         
-
-    //     foreach ($data as $record) {
-
-    //         $record = $this->formatChecker($record);
-
-    //         if ($record->category === 'kid') {
-
-    //             $record = $this->formatChecker($record);
-    //             $image = Image::where('id', $record->id)->first();
-    //             if($image){
-    //                 $storagePath = $image->path;
-    //                 $url = url('storage/' . $storagePath);
-    //                 $kidArray[]['images'] = [
-    //                     'id' => $image->id,
-    //                     'imageUrl' => $url
-    //                 ];
-    //             }
-
-    //             $kidArray[] = $record;
-    //         } elseif ($record->category === 'pet') {
-    //             $petArray[] = $record;
-    //         } else {
-    //             $luggageArray[] = $record;
-    //         }
-    //     }
-
-    //     $combinedData = [
-    //         "kid" => $kidArray,
-    //         "pet" => $petArray,
-    //         "luggage" => $luggageArray
-    //     ];
-
-    //     return response()->json(
-    //         [
-    //             "status" => true,
-    //             'message' => "Data successfully fetched",
-    //             'data' => $combinedData
-    //         ],
-    //         200
-    //     );
-    // }
-
-    public function getAllTags()
-{
+    public function getAllTags(){
     $data = Tag::all();
 
     $kidArray = [];
@@ -330,6 +280,8 @@ class TagController extends Controller
     {
         $user = Auth::user();
         $userId = $user->id;
+
+        // return $request;
 
 
         $validator = Validator::make($request->all(), [
@@ -428,8 +380,19 @@ class TagController extends Controller
 
     public function deleteTag($id)
     {
-        $tag = Tag::find($id);
+        $user = Auth::user();
+        if(!$user){
+            return response()->json(["status"=>false, "message"=>"Unauthorized"],401);
+        }
+        $tag = Tag::where('id', $id)
+        ->where('userId', $user->id)
+        ->first();
         if ($tag) {
+            $tagImages = Image::where("tagId", $tag->id)->get();
+            foreach($tagImages as $image){
+                Storage::delete('public/uploads/'. basename($image->path));
+                $image->delete();
+            }
             $tag->delete();
             return response()->json([
                 "status" => true,
